Make packages with native dependencies optional
These packages are only used on device implementations and so
aren't typically used in development scenarios. They cause build
errors that prevent contributors from easily getting involved
so we mark them optional. Obviously they are not optional in deployment
but they *will* be installed if possible and it's our responsibility
to ensure that the OS's in deployment have the correct native libraries.