remove CSS code duplication (popular <---> WOS)
I'm not sure if it's kosher to share CSS between "apps" like this,
but the popular CSS was mostly identical to the WOS CSS. To remove the
need for us to manually update multiple places when changing the CSS
(e.g. to fix the title centering issue I just fixed), it's nice to have
this code in one place where possible.