Abstract WOS CSS' #title decl to .main-title; sty
Removes the need for all of these unrelated apps to depend on the WOS
CSS -- now they can just refer to the terminal CSS. This is nice,
because it means that if we need to change the WOS CSS in the future
then it won't have the risk of impacting unrelated stuff. (With the
exception of the "popular" app, which still depends on other stuff
from the WOS CSS, but that's a whole other thing.)

Also, a few other styling things -- added space between <p> links
on the index page, added padding to the history pages, etc.